Moved wkt from root dir to traveltimepy
