--- wkt/__init__.py
@@ -1,1 +1,0 @@
-"""WKT wrapper for Shapely for working with WKT type objects"""
