--- wkt/tests/__init__.py
@@ -1,1 +1,0 @@
-"""WKT tests"""
